requirements.txt
py-clob-client
python-dotenv
requests
apscheduler
yfinance
scipy
xgboost
pandas
scikit-learn
